+++ src/PortalBase/CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Change `\Heptacom\HeptaConnect\Portal\Base\Support\Contract\DeepObjectIteratorContract::iterate` to resolve nested iterables deferred instead of in the moment they are found to lower memory peaks during larger object inspection
+
 ### Deprecated
 
 ### Removed